# openapi.yaml
openapi: 3.1.1
info:
  title: Blog API
  version: 1.0.0
  description: API REST para gestionar autores y posts de un blog
  contact:
    name: Equipo de Desarrollo
    email: dev@blog.com

servers:
  - url: http://localhost:3000
    description: Desarrollo local
  - url: https://blog-api-production.up.railway.app
    description: Producción

paths:
  /api/authors:
    get:
      summary: Obtener todos los autores
      description: Devuelve la lista completa de autores ordenados por nombre
      tags:
        - Autores
      responses:
        '200':
          description: Lista de autores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Author'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Crear un nuevo autor
      tags:
        - Autores
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
              properties:
                name:
                  type: string
                  description: Nombre completo del autor
                email:
                  type: string
                  format: email
                  description: Email del autor (debe ser único)
                bio:
                  type: string
                  description: Biografía del autor (opcional)
      responses:
        '201':
          description: Autor creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email ya existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/authors/{id}:
    get:
      summary: Obtener un autor por ID
      tags:
        - Autores
      parameters:
        - name: id
          in: path
          required: true
          description: ID del autor
          schema:
            type: integer
      responses:
        '200':
          description: Autor encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '404':
          description: Autor no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Author:
      type: object
      properties:
        id:
          type: integer
          description: ID único del autor
        name:
          type: string
          description: Nombre completo del autor
        email:
          type: string
          format: email
          description: Email del autor
        bio:
          type: string
          description: Biografía del autor
        created_at:
          type: string
          format: date-time
          description: Fecha de creación   
    Error:
      type: object
      properties:
        error:
          type: string
          description: Mensaje de error